% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minPath.R
\name{distBoot}
\alias{distBoot}
\title{Compute a bootstrap distribution for testing minimal path lengths}
\usage{
distBoot(x, partial = FALSE, unsigned = TRUE, output = "orig")
}
\arguments{
\item{x}{A data matrix}

\item{partial}{Identical to the "partial" parameter of pMatrix}

\item{unsigned}{Identical to the "unsigned" parameter of pMatrix}

\item{output}{Identical to the "output" parameter of pMatrix}
}
\value{
A vector of path lengths obtained from a permuted data matrix
}
\description{
Compute a bootstrap distribution for testing minimal path lengths
}
\examples{
db <- distBoot(foldChange, partial=TRUE)
}
