% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFuncs.R
\name{dLaplace}
\alias{dLaplace}
\title{Computes a p-value matrix from the difference of two p-value matrices.}
\usage{
dLaplace(pmat1, pmat2, twoTailed = FALSE, unifOut = FALSE)
}
\arguments{
\item{pmat1}{First p-value matrix}

\item{pmat2}{Second p-value matrix}

\item{twoTailed}{Compute a two-tailed p-value}

\item{unifOut}{Force the output to a uniform distribution}
}
\value{
A new p-value matrix testing the ratio of the first p-value matrix
over the second (or vice versa if twoTailed = TRUE).
}
\description{
Computes a p-value matrix from the difference of two p-value matrices.
}
\examples{
pm1 <- pMatrix(foldChange)
pm2 <- pMatrix(foldChange, z=traitData["TCDD",])
lapMat <- dLaplace(pm1, pm2)
}
