% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gFuncs.R
\name{eigenGenes}
\alias{eigenGenes}
\title{Compute eigengenes following gamma clustering}
\usage{
eigenGenes(x, m)
}
\arguments{
\item{x}{A data matrix}

\item{m}{Membership probabilities}
}
\value{
A matrix of eigengenes where rows correspond to each cluster.
}
\description{
Compute eigengenes following gamma clustering
}
\examples{
eigenGenes(foldChange, clusters$membership)
}
