% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{gammaCluster}
\alias{gammaCluster}
\title{Perform gamma clustering to identify clusters that maximize the probability
of null inter-cluster correlations}
\usage{
gammaCluster(pmat, maxiter = 100)
}
\arguments{
\item{pmat}{A p-value matrix}

\item{maxiter}{Maximum number of iterations}
}
\value{
A list containing cluster labels and membership probabilities
}
\description{
Perform gamma clustering to identify clusters that maximize the probability
of null inter-cluster correlations
}
\examples{
pm <- pMatrix(foldChange)
clusters <- gammaCluster(pm)
}
