% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{gammaPlot}
\alias{gammaPlot}
\title{Visualize clusters with UMAP}
\usage{
gammaPlot(pmat, labels)
}
\arguments{
\item{pmat}{A p-value matrix}

\item{labels}{Cluster labels to visualize}
}
\description{
Visualize clusters with UMAP
}
\examples{
pm <- pMatrix(foldChange)
clusters <- gammaCluster(pm)
gammaPlot(pm, clusters$labels)
}
