% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minPath.R
\name{minPath}
\alias{minPath}
\title{Calculate a path between a source and target with the minimum likelihood
of observing any links under the null hypothesis}
\usage{
minPath(s, t, pmat, r = 1, clusters = NULL, nodeNames = NULL, plotpath = TRUE)
}
\arguments{
\item{s}{Name (or index) of the source node}

\item{t}{Name (or index) of the target node}

\item{pmat}{A p-value matrix}

\item{r}{A regularization constant, the log of which will be added to all
distances}

\item{clusters}{A vector of cluster labels. If null, all nodes will be
the same color}

\item{nodeNames}{A vector of node names. If null, indices will be used instead}

\item{plotpath}{If TRUE, plot the resulting path}
}
\value{
The minimal path connecting the source to target
}
\description{
Calculate a path between a source and target with the minimum likelihood
of observing any links under the null hypothesis
}
\examples{
pm <- pMatrix(foldChange, traitData, unsigned=TRUE, output="orig")
p <- minPath(1, 2, pm, r=1-1/dim(foldChange)[1])
}
