% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixCalcs.R
\name{pMatrix}
\alias{pMatrix}
\title{Compute a p-value matrix}
\usage{
pMatrix(
  features,
  traits = NULL,
  z = NULL,
  unsigned = FALSE,
  partial = FALSE,
  output = "unif"
)
}
\arguments{
\item{features}{A data matrix of features. Rows are features.}

\item{traits}{A data matrix of traits. Rows are traits.}

\item{z}{A matrix or vector of covariates for conditioning correlations}

\item{unsigned}{If TRUE, ignore the sign of correlations}

\item{partial}{If TRUE, compute a partial correlation matrix}

\item{output}{One of "unif", "orig", or "cor" corresponding to forced-uniform,
original p-values, or raw correlations, respectively}
}
\value{
A (N x N) matrix of p-values. The first rows/columns correspond to
traits if supplied
}
\description{
Compute a p-value matrix
}
\examples{
pm <- pMatrix(foldChange)
}
