% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{recGamma}
\alias{recGamma}
\title{Gamma sub-clustering performed on each individual cluster before a final
optimization step. The output can be re-entered as input for recursive
optimization.}
\usage{
recGamma(pmat, labels, maxiter = 10)
}
\arguments{
\item{pmat}{A p-value matrix}

\item{labels}{A vector of previous cluster labels}

\item{maxiter}{Maximum number of initial splits of each cluster}
}
\value{
A list of cluster labels and membership probabilities
}
\description{
Gamma sub-clustering performed on each individual cluster before a final
optimization step. The output can be re-entered as input for recursive
optimization.
}
\examples{
pm <- pMatrix(foldChange)
clusters <- gammaCluster(pm)
rcluster <- recGamma(pmat, clusters$labels)
rcluster <- recGamma(pmat, rcluster$labels)
}
